requirements.txt
beautifulsoup4>=4.12.0
html2text>=2020.1.16
openai>=1.3.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
slack-sdk>=3.26.0
pydantic
beautifulsoup4